convert round to np.int

+++ deepchem/models/xgboost_models/__init__.py
@@ -33,5 +33,5 @@ class XGBoostModel(SklearnModel):
     # Since test size is 20%, when retrain model to whole data, expect
     # n_estimator increased to 1/0.8 = 1.25 time.
     estimated_best_round = np.round(self.model_instance.best_ntree_limit * 1.25)
-    self.model_instance.n_estimators = estimated_best_round
+    self.model_instance.n_estimators = np.int64(estimated_best_round)
     self.model_instance.fit(X_train, y_train, eval_metric="auc", verbose=False)
